requirement.txt
openai-whisper
moviepy 
ffmpeg-python
